requirements.txt
Flask
Flask-WTF
Flask-Assets
Flask-Script
Flask-Mail
Flask-SocketIO
Werkzeug
jsmin
cssmin
pyscss
Jinja2
SQLAlchemy
alembic
psycopg2
PyYAML
Pygments
passlib
pytz
py-bcrypt
requests
celery
librabbitmq
kombu
rfc3987
sphinx
mock
python-magic
Pillow
pytest

